requirements.txt
pyrogram
tgcrypto
python-dotenv
apscheduler